requirements.txt
pandas>=2.2.1
google-api-python-client>=2.121.0
google.generativeai>=0.4.0